move more vault utils to e2e utils
