+++ e2e/examples/eth_eureka_vault/vault.rs
@@ -1,36 +1,41 @@
-use std::{error::Error, str::FromStr, thread, time::Duration};
+use std::{
+    error::Error,
+    str::FromStr,
+    time::{Duration, SystemTime},
+};
 
 use alloy::primitives::Address;
 
+use evm::setup_eth_accounts;
 use localic_utils::{
-    ConfigChainBuilder, TestContextBuilder, LOCAL_IC_API_URL, NEUTRON_CHAIN_ADMIN_ADDR,
-    NEUTRON_CHAIN_ID,
+    utils::ethereum::EthClient, ConfigChainBuilder, TestContextBuilder, LOCAL_IC_API_URL,
+    NEUTRON_CHAIN_ADMIN_ADDR, NEUTRON_CHAIN_ID,
 };
 
 use log::{info, warn};
 
+use program::{setup_neutron_libraries, upload_neutron_contracts};
 use valence_chain_client_utils::{
     cosmos::base_client::BaseClient,
     evm::{base_client::EvmBaseClient, request_provider_client::RequestProviderClient},
     neutron::NeutronClient,
 };
 
+mod program;
+
 use valence_e2e::{
     async_run,
     utils::{
         astroport::setup_astroport_cl_pool,
         authorization::set_up_authorization_and_processor,
-        ethereum::{
-            self as ethereum_utils, set_up_anvil_container, ANVIL_NAME, DEFAULT_ANVIL_PORT,
-        },
+        ethereum::{set_up_anvil_container, ANVIL_NAME, DEFAULT_ANVIL_PORT},
         parse::{get_chain_field_from_local_ic_log, get_grpc_address_and_port_from_url},
         solidity_contracts::{BaseAccount, MockERC20},
-        DEFAULT_ANVIL_RPC_ENDPOINT, ETHEREUM_CHAIN_NAME, LOGS_FILE_PATH, VALENCE_ARTIFACTS_PATH,
+        DEFAULT_ANVIL_RPC_ENDPOINT, LOGS_FILE_PATH, VALENCE_ARTIFACTS_PATH,
     },
 };
 
 const ASTROPORT_CONCENTRATED_PAIR_TYPE: &str = "concentrated";
-const VAULT_NEUTRON_CACHE_PATH: &str = "e2e/examples/eth_vault/neutron_contracts/";
 
 mod evm;
 mod strategist;
@@ -40,6 +45,7 @@ const WBTC_WHALE: &str = "0x0555E30da8f98308EdB960aa94C0Db47230d2B9c";
 const EUREKA_HANDLER: &str = "0xfc2d0487a0ae42ae7329a80dc269916a9184cf7c";
 const EUREKA_HANDLER_SRC_CLIENT: &str = "cosmoshub-0";
 const WBTC_NEUTRON_DENOM: &str = "WBTC";
+const VAULT_NEUTRON_CACHE_PATH: &str = "e2e/examples/eth_eureka_vault/neutron_contracts/";
 
 // #[tokio::main]
 fn main() -> Result<(), Box<dyn Error>> {
@@ -62,9 +68,14 @@ fn main() -> Result<(), Box<dyn Error>> {
     )
     .unwrap();
 
+    let eth = EthClient::new(DEFAULT_ANVIL_RPC_ENDPOINT)?;
+
     let eth_rp = async_run!(rt, eth_client.get_request_provider().await.unwrap());
 
     let eth_accounts = async_run!(rt, eth_client.get_provider_accounts().await.unwrap());
+    let eth_admin_acc = eth_accounts[0];
+
+    let ethereum_program_accounts = setup_eth_accounts(&rt, &eth_client, eth_admin_acc)?;
 
     let eth_admin_acc = eth_accounts[0];
     let _eth_user_acc = eth_accounts[2];
@@ -79,32 +90,6 @@ fn main() -> Result<(), Box<dyn Error>> {
     );
 
     info!("admin balance: {:?}", admin_balance);
-
-    let deposit_account_init_tx =
-        BaseAccount::deploy_builder(&eth_rp, eth_admin_acc, vec![]).into_transaction_request();
-    let withdraw_account_init_tx =
-        BaseAccount::deploy_builder(&eth_rp, eth_admin_acc, vec![]).into_transaction_request();
-
-    let deposit_account_rx = async_run!(
-        rt,
-        eth_client
-            .execute_tx(deposit_account_init_tx.clone())
-            .await
-            .unwrap()
-    );
-    let withdraw_account_rx = async_run!(
-        rt,
-        eth_client
-            .execute_tx(withdraw_account_init_tx.clone())
-            .await
-            .unwrap()
-    );
-
-    let deposit_account_addr = deposit_account_rx.contract_address.unwrap();
-    let withdraw_account_addr = withdraw_account_rx.contract_address.unwrap();
-
-    info!("deposit account address: {deposit_account_addr}");
-    info!("withdraw account address: {withdraw_account_addr}");
 
     let wbtc_token_address = Address::from_str(WBTC_ERC20).unwrap();
     let eureka_handler_address = Address::from_str(EUREKA_HANDLER).unwrap();
@@ -188,6 +173,36 @@ fn main() -> Result<(), Box<dyn Error>> {
         setup_astroport_cl_pool(&mut test_ctx, wbtc_on_neutron.to_string())?;
 
     info!("BTC-NTRN cl pool: {:?}", pool_addr);
+
+    // set up the authorization and processor contracts on neutron
+    let (authorization_contract_address, neutron_processor_address) =
+        set_up_authorization_and_processor(
+            &mut test_ctx,
+            hex::encode(
+                SystemTime::now()
+                    .duration_since(SystemTime::UNIX_EPOCH)?
+                    .as_secs()
+                    .to_string(),
+            ),
+        )?;
+
+    upload_neutron_contracts(&mut test_ctx)?;
+
+    let neutron_program_accounts = program::setup_neutron_accounts(&mut test_ctx)?;
+
+    let amount_to_transfer = 1_000_000;
+
+    let neutron_program_libraries = setup_neutron_libraries(
+        &mut test_ctx,
+        &neutron_program_accounts,
+        &pool_addr,
+        &authorization_contract_address,
+        &neutron_processor_address,
+        amount_to_transfer,
+        &wbtc_on_neutron,
+        ethereum_program_accounts.withdraw.to_string(),
+        &lp_token,
+    )?;
 
     rt.shutdown_background();
