+++ e2e/src/utils/astroport.rs
@@ -9,13 +9,19 @@ use localic_utils::{
 
 use log::info;
 use serde_json::Value;
+use valence_astroport_lper::msg::LiquidityProviderConfig;
 use valence_astroport_utils::astroport_native_lp_token::{
     Asset, AssetInfo, ConcentratedLiquidityExecuteMsg, ConcentratedPoolParams,
     FactoryInstantiateMsg, FactoryQueryMsg, NativeCoinRegistryExecuteMsg,
     NativeCoinRegistryInstantiateMsg, PairConfig, PairType,
 };
+use valence_library_utils::{liquidity_utils::AssetData, LibraryAccountType};
 
-use crate::utils::{ASTROPORT_PATH, GAS_FLAGS, LOCAL_CODE_ID_CACHE_PATH_NEUTRON};
+use crate::utils::{
+    base_account::approve_library,
+    manager::{ASTROPORT_LPER_NAME, ASTROPORT_WITHDRAWER_NAME},
+    ASTROPORT_PATH, GAS_FLAGS, LOCAL_CODE_ID_CACHE_PATH_NEUTRON,
+};
 
 const _PROVIDE_LIQUIDITY_AUTHORIZATIONS_LABEL: &str = "provide_liquidity";
 const _WITHDRAW_LIQUIDITY_AUTHORIZATIONS_LABEL: &str = "withdraw_liquidity";
@@ -266,4 +272,136 @@ pub fn setup_astroport_cl_pool(
     std::thread::sleep(std::time::Duration::from_secs(3));
 
     Ok((pool_addr.to_string(), lp_token.to_string()))
+}
+
+pub fn setup_astroport_lper_lib(
+    test_ctx: &mut TestContext,
+    input_account: String,
+    output_account: String,
+    asset_data: AssetData,
+    pool_addr: String,
+    _processor: String,
+    _authorizations: String,
+) -> Result<String, Box<dyn Error>> {
+    let lper_code_id = test_ctx
+        .get_contract()
+        .contract(ASTROPORT_LPER_NAME)
+        .get_cw()
+        .code_id
+        .unwrap();
+
+    let astro_cl_pair_type = valence_astroport_utils::astroport_native_lp_token::PairType::Custom(
+        ASTROPORT_CONCENTRATED_PAIR_TYPE.to_string(),
+    );
+
+    let astro_lp_config = LiquidityProviderConfig {
+        pool_type: valence_astroport_utils::PoolType::NativeLpToken(astro_cl_pair_type.clone()),
+        asset_data,
+        max_spread: None,
+    };
+
+    let astro_lper_library_cfg = valence_astroport_lper::msg::LibraryConfig {
+        input_addr: LibraryAccountType::Addr(input_account.to_string()),
+        output_addr: LibraryAccountType::Addr(output_account.to_string()),
+        lp_config: astro_lp_config,
+        pool_addr,
+    };
+
+    let astroport_lper_instantiate_msg =
+        valence_library_utils::msg::InstantiateMsg::<valence_astroport_lper::msg::LibraryConfig> {
+            // TODO: uncomment to not bypass authorizations/processor logic
+            // owner: authorizations.to_string(),
+            // processor: processor.to_string(),
+            owner: NEUTRON_CHAIN_ADMIN_ADDR.to_string(),
+            processor: NEUTRON_CHAIN_ADMIN_ADDR.to_string(),
+            config: astro_lper_library_cfg,
+        };
+
+    let astro_lper_lib = contract_instantiate(
+        test_ctx
+            .get_request_builder()
+            .get_request_builder(NEUTRON_CHAIN_NAME),
+        DEFAULT_KEY,
+        lper_code_id,
+        &serde_json::to_string(&astroport_lper_instantiate_msg)?,
+        "astro_lper",
+        None,
+        "",
+    )?;
+    info!("astro lper lib: {}", astro_lper_lib.address);
+
+    info!("approving astro lper library on deposit account...");
+    approve_library(
+        test_ctx,
+        NEUTRON_CHAIN_NAME,
+        DEFAULT_KEY,
+        &input_account,
+        astro_lper_lib.address.to_string(),
+        None,
+    );
+
+    Ok(astro_lper_lib.address)
+}
+
+pub fn setup_astroport_lwer_lib(
+    test_ctx: &mut TestContext,
+    input_account: String,
+    output_account: String,
+    asset_data: AssetData,
+    pool_addr: String,
+    _processor: String,
+) -> Result<String, Box<dyn Error>> {
+    let lwer_code_id = test_ctx
+        .get_contract()
+        .contract(ASTROPORT_WITHDRAWER_NAME)
+        .get_cw()
+        .code_id
+        .unwrap();
+
+    let astro_cl_pair_type = valence_astroport_utils::astroport_native_lp_token::PairType::Custom(
+        ASTROPORT_CONCENTRATED_PAIR_TYPE.to_string(),
+    );
+
+    let astro_lw_config = valence_astroport_withdrawer::msg::LiquidityWithdrawerConfig {
+        pool_type: valence_astroport_utils::PoolType::NativeLpToken(astro_cl_pair_type),
+        asset_data,
+    };
+    let astro_lwer_library_cfg = valence_astroport_withdrawer::msg::LibraryConfig {
+        input_addr: LibraryAccountType::Addr(input_account.to_string()),
+        output_addr: LibraryAccountType::Addr(output_account.to_string()),
+        withdrawer_config: astro_lw_config,
+        pool_addr: pool_addr.to_string(),
+    };
+    let astroport_lwer_instantiate_msg = valence_library_utils::msg::InstantiateMsg::<
+        valence_astroport_withdrawer::msg::LibraryConfig,
+    > {
+        owner: NEUTRON_CHAIN_ADMIN_ADDR.to_string(),
+        processor: NEUTRON_CHAIN_ADMIN_ADDR.to_string(),
+        config: astro_lwer_library_cfg,
+    };
+
+    let astro_lwer_lib = contract_instantiate(
+        test_ctx
+            .get_request_builder()
+            .get_request_builder(NEUTRON_CHAIN_NAME),
+        DEFAULT_KEY,
+        lwer_code_id,
+        &serde_json::to_string(&astroport_lwer_instantiate_msg)?,
+        "astro_lwer",
+        None,
+        "",
+    )?;
+    info!("astro lwer lib: {}", astro_lwer_lib.address);
+
+    info!("approving astro lwer library on position account...");
+    approve_library(
+        test_ctx,
+        NEUTRON_CHAIN_NAME,
+        DEFAULT_KEY,
+        &input_account,
+        astro_lwer_lib.address.to_string(),
+        None,
+    );
+
+    Ok(astro_lwer_lib.address)
 }
